requirements/base.txt
Click==7.0
click==7.1.2
wiktionaryparser==0.0.97
progressbar2==3.53.1
